Long to Int

+++ Kendaraan.java
@@ -1,10 +1,10 @@
 public abstract class Kendaraan {
-    protected long id;
+    protected int id;
     protected String nama;
     protected FaktorEmisi emisi;
     protected double efisiensiKmPerLiter;
 
-    public Kendaraan(long id, String nama, FaktorEmisi emisi, double efisiensiKmPerLiter) {
+    public Kendaraan(int id, String nama, FaktorEmisi emisi, double efisiensiKmPerLiter) {
         this.id = id;
         this.nama = nama;
         this.emisi = emisi;
